+++ js/data.js
@@ -1,0 +1,34 @@
+export const postsCount = 25;
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+
+export const DESCRIPTIONS = [
+  'Прекрасный закат на море',
+  'Моя пушистая кошка',
+  'Горный поход выходного дня',
+  'Кофе в уютном кафе',
+  'Прогулка по осеннему парку',
+  'Архитектура старого города',
+  'Мой завтрак',
+  'Вечерний город',
+  'Природа в её лучшем проявлении',
+  'Момент с друзьями'
+];
